class化してリファクタリング
