+++ .gitignore
@@ -1,13 +1,13 @@
 # Python
-/py/__pycache__/
-/py/*.pyc
-/py/*.pyo
-/py/*.pyd
-/py/*.log
-/py/build/*
-!/py/build/.gitkeep
-/py/dist/*
-!/py/dist/.gitkeep
+/py/*/__pycache__/
+/py/*/*.pyc
+/py/*/*.pyo
+/py/*/*.pyd
+/py/*/*.log
+/py/*/build/*
+!/py/*/build/.gitkeep
+/py/*/dist/*
+!/py/*/dist/.gitkeep
 
 # files
 /py/*/*.csv
